requirements.txt
paho-mqtt>=2.0.0
flask
loguru
python-dotenv